Split packet definitions out from parser.cs Added a few known/half known model 3 packets
