+++ CANBUSAnalyzer/CANBUSAnalyzer.csproj
@@ -96,6 +96,8 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="BindableTwoDArray.cs" />
+    <Compile Include="Model3Packets.cs" />
+    <Compile Include="ModelSPackets.cs" />
     <Compile Include="ObservableDictionary.cs" />
     <Compile Include="Packet.cs" />
     <Compile Include="Parser.cs" />
